move testdata to internal
