+++ internal/testdata/hammer.cfg
